refactor: rename api directory to web
