+++ src/main/java/nextstep/web/GlobalControllerAdvice.java
@@ -1,6 +1,6 @@
-package nextstep.api;
+package nextstep.web;
 
-import nextstep.api.dto.ErrorResponse;
+import nextstep.web.dto.ErrorResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
