+++ src/main/java/nextstep/web/reservation/ReservationController.java
@@ -1,7 +1,7 @@
-package nextstep.api.reservation;
+package nextstep.web.reservation;
 
-import nextstep.api.reservation.dto.ReservationCreateRequest;
-import nextstep.api.reservation.dto.ReservationResponse;
+import nextstep.web.reservation.dto.ReservationCreateRequest;
+import nextstep.web.reservation.dto.ReservationResponse;
 import nextstep.domain.reservation.service.ReservationService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
